+++ application_sandbox/external_image/image_import.cpp
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <chrono>
+
 #include "application_sandbox/sample_application_framework/sample_application.h"
+#include "mathfu/matrix.h"
+#include "mathfu/vector.h"
 #include "support/entry/entry.h"
 #include "vulkan_helpers/buffer_frame_data.h"
 #include "vulkan_helpers/helper_functions.h"
 #include "vulkan_helpers/vulkan_application.h"
 #include "vulkan_helpers/vulkan_model.h"
 #include "vulkan_helpers/vulkan_texture.h"
-
-#include <chrono>
-#include "mathfu/matrix.h"
-#include "mathfu/vector.h"
 
 #ifdef __linux__
 #include <sys/socket.h>
@@ -561,7 +561,8 @@ class TexturedCubeSample
     } while (pipe_handle == INVALID_HANDLE_VALUE);
     DWORD bytes_read;
     ReadFile(pipe_handle, native_handles.data(),
-             sizeof(HANDLE) * native_handles.size(), &bytes_read, NULL);
+             static_cast<DWORD>(sizeof(HANDLE) * native_handles.size()),
+             &bytes_read, NULL);
 
     CloseHandle(pipe_handle);
   }
